first wave of fixes
