+++ apps/web/src/types/index.ts
@@ -1,5 +1,6 @@
 export * from './color'
 export * from './conditionOperators'
+export * from './companyDataType'
 export * from './elementsPropsTypes'
 export * from './formTypes'
 export * from './genericTypes'
@@ -15,6 +16,5 @@ export * from './track/trackType'
 export * from './translationTypes'
 export * from './utilityTypes'
 export * from './breakpointsTypes'
-export * from './identification'
 export * from '@tee/common'
 export * from '@tee/data'
